+++ creditcards/semi/form-constant.js
@@ -1,0 +1,14 @@
+const semiFormContant = {
+  // SEMI
+  formPath: ['semi', 'smart-emi', 'smart emi', 'smart_emi', 'smartemi'],
+  class: 'semi-form',
+  urlKey: ['semi', 'smart-emi', 'smart emi', 'smartemi'],
+  launchScript: {
+    dev: 'https://assets.adobedtm.com/80673311e435/029b16140ccd/launch-94203efd95a9-staging.min.js',
+    prod: 'https://assets.adobedtm.com/80673311e435/029b16140ccd/launch-39d52f236cd6.min.js',
+    loadTime: 1200,
+  },
+  stylePath: '/creditcards/semi/styles/semi.css',
+};
+
+export default semiFormContant;
