+++ styles/styles.css
@@ -16,7 +16,7 @@
  @import url('./header-banner.css');
  @import url('./nre-nro.css');
  @import url('./fd-styles.css');
- @import url('./semi-styles.css');
+ @import url('../creditcards/semi/styles/semi-styles.css');
  @import url('./accordian-panel-styles.css');
 
  :root {
